Add files via upload
This is my first web page that I have tried but it includes some error like, the web page that I have made looks perfect only in Microsoft edge, and if opened in any other browser it changes somewhat. And mobile view is also not supported. 😰
vedantcreation.000webhostapp.com   this is the web link to my web page. if anyone could help me sorting out this problem, I would be really happy, it would be a great help for me.